update support-faq screen structure
